Merge branch 'master' of https://github.com/klimesf/a4m36jee-project into jsf

# Conflicts:
#	README.md
#	src/main/java/cz/cvut/fel/a4m36jee/airlines/model/Flight.java
#	src/main/java/cz/cvut/fel/a4m36jee/airlines/model/Reservation.java
#	src/main/java/cz/cvut/fel/a4m36jee/airlines/service/DestinationService.java
#	src/main/java/cz/cvut/fel/a4m36jee/airlines/service/FlightService.java
#	src/main/java/cz/cvut/fel/a4m36jee/airlines/service/ReservationService.java
#	src/main/webapp/WEB-INF/web.xml
#	src/test/java/cz/cvut/fel/a4m36jee/airlines/rest/ReservationResourceTest.java

+++ src/main/java/cz/cvut/fel/a4m36jee/airlines/view/util/reservation/PasswordJSFValidator.java
@@ -31,7 +31,7 @@ public class PasswordJSFValidator implements Validator {
         }
         String pass = (String) o;
         String idString = (String) uiComponent.getAttributes().get("reservationId");
-        Integer id = Integer.parseInt(idString);
+        Long id = Long.parseLong(idString);
         if(id == null) {
             FacesMessage facesMsg = new FacesMessage("Reservation not found!");
             facesMsg.setSeverity(FacesMessage.SEVERITY_ERROR);
